+++ test/bot_test.dart
@@ -9,11 +9,24 @@ const controllerPort = '/dev/ttyACM0';
 /// Configuration directory containing the manufacturer_specific.xml file.
 const configPath = '/usr/local/etc/openzwave/';
 
-main() async {
+main(List<String> args) async {
+  bool waitForAllUpdated = false;
+  Timeout timeout = new Timeout(new Duration(minutes: 1));
+
+  int argIndex = 0;
+  while (argIndex < args.length) {
+    if (args[argIndex] == '--waitForAllUpdated') {
+      waitForAllUpdated = true;
+    } else if (args[argIndex].startsWith('--timeout=')) {
+      timeout = new Timeout(
+          new Duration(minutes: int.parse(args[argIndex].substring(10))));
+    } else {
+      print('unknown argument: ${args[argIndex]}');
+    }
+    ++argIndex;
+  }
+
   ZWave zwave;
-  tearDown(() async {
-    await zwave?.dispose();
-  });
   test('bot', () async {
     print('initializing...');
     zwave = await ZWave.init(configPath);
@@ -24,12 +37,17 @@ main() async {
     await zwave.connect(controllerPort);
     print('updating...');
     await zwave.update();
+    if (waitForAllUpdated) {
+      print('waiting for all nodes to respond...');
+      await zwave.allUpdated();
+    }
     for (Device device in zwave.devices) {
       print('  $device');
-      print('    node basic    ${device.nodeBasic}');
-      print('    node generic  ${device.nodeGeneric}');
-      print('    node specific ${device.nodeSpecific}');
-      print('    node type     ${device.nodeType}');
+      print('    ${device.nodeBasic} ${device.nodeGeneric}'
+          ' ${device.nodeSpecific} ${device.nodeType}');
+      print('    ${device.manufacturerId} ${device.manufacturerName}');
+      print('    ${device.productId} ${device.productType}'
+          ' ${device.productName}');
       for (Value value in device.values) {
         var buf = new StringBuffer('      $value');
         while (buf.length < 50) buf.write(' ');
@@ -53,5 +71,9 @@ main() async {
     }
     print('disposing...');
     await zwave.dispose();
+    zwave = null;
+  }, timeout: timeout);
+  tearDown(() async {
+    await zwave?.dispose();
   });
 }