+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.1.3
+* add device accessors for product and manufacturer information
+* write network config to file on shutdown
+* cleanup generated code and usages
+
 ## 0.1.2
 * build device value list from notifications